% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_trace}
\alias{example_trace}
\title{Example trace}
\format{
A tibble with 200 rows and 2 variables:
\describe{
\item{timestamp}{POSIXct timestamps of observations}
\item{location}{Locations of observations for one of four places (Rivendell, Tatooine, Shangri-La, Narnia)}
}
}
\usage{
example_trace
}
\description{
A dataset containing trace of a fictional migrant in the form of location - timestamp pairs.
}
\keyword{datasets}
